--- src/components/cards/categoryCardTypes.type.tsx
@@ -1,6 +1,0 @@
-export type categoryCardTypes = {
-	id: number;
-	img: string;
-	title: string;
-	categoryName: string;
-};
